requirements.txt

# Core ML libraries
pandas>=2.0.3
numpy>=1.26.4
scikit-learn>=1.3.2
joblib>=1.3.2
xgboost>=2.0.0
pyyaml>=6.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.2

# Configuration
python-dotenv>=1.0.0

# AI/LLM
google-genai>=1.0.0

# CLI formatting
rich>=13.6.0
